Refactor!: split components into numerous ui submodules
